+++ werkbank/Properties/Resources.Designer.cs
@@ -73,6 +73,16 @@ namespace werkbank.Properties {
         /// <summary>
         ///   Looks up a localized resource of type System.Drawing.Bitmap.
         /// </summary>
+        internal static System.Drawing.Bitmap btn_copy {
+            get {
+                object obj = ResourceManager.GetObject("btn_copy", resourceCulture);
+                return ((System.Drawing.Bitmap)(obj));
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized resource of type System.Drawing.Bitmap.
+        /// </summary>
         internal static System.Drawing.Bitmap btn_create {
             get {
                 object obj = ResourceManager.GetObject("btn_create", resourceCulture);