+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 [v3.5.0](https://github.com/sedlatschek/werkbank3/releases/tag/v3.5.0.0)
 - fix bug where werk would disappear after a transition, if Werkbank was restarted after the delete operation was done
+- add button to copy queue operation errors
 
 [v3.4.0](https://github.com/sedlatschek/werkbank3/releases/tag/v3.4.0.0)
 - fix bug where zip operation would skip empty directories